requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
requests==2.31.0
google-generativeai==0.4.1
